flask/requirements.txt
pymongo
flask
pyyaml
gunicorn